moving some classes around
