+++ src/main/java/com/controllerface/bvge/cl/GPUCoreMemory.java
@@ -9,7 +9,8 @@ import com.controllerface.bvge.cl.kernels.crud.*;
 import com.controllerface.bvge.cl.programs.GPUCrud;
 import com.controllerface.bvge.cl.programs.GPUProgram;
 import com.controllerface.bvge.cl.programs.ScanDeletes;
-import com.controllerface.bvge.ecs.systems.sectors.*;
+import com.controllerface.bvge.game.world.sectors.SectorContainer;
+import com.controllerface.bvge.game.world.sectors.*;
 import com.controllerface.bvge.geometry.ModelRegistry;
 import com.controllerface.bvge.physics.PhysicsEntityBatch;
 import com.controllerface.bvge.physics.PhysicsObjects;
