--- src/main/java/com/controllerface/bvge/ecs/systems/sectors/Sector.java
@@ -1,3 +1,0 @@
-package com.controllerface.bvge.ecs.systems.sectors;
-
-public record Sector(int x, int y) { }
