+++ src/main/java/com/controllerface/bvge/game/TestGame.java
@@ -5,10 +5,10 @@ import com.controllerface.bvge.ecs.ECS;
 import com.controllerface.bvge.ecs.components.*;
 import com.controllerface.bvge.ecs.systems.CameraTracking;
 import com.controllerface.bvge.ecs.systems.GameSystem;
-import com.controllerface.bvge.ecs.systems.sectors.InventorySystem;
-import com.controllerface.bvge.ecs.systems.sectors.Sector;
-import com.controllerface.bvge.ecs.systems.sectors.WorldLoader;
-import com.controllerface.bvge.ecs.systems.sectors.WorldUnloader;
+import com.controllerface.bvge.ecs.systems.InventorySystem;
+import com.controllerface.bvge.game.world.sectors.Sector;
+import com.controllerface.bvge.game.world.WorldLoader;
+import com.controllerface.bvge.game.world.WorldUnloader;
 import com.controllerface.bvge.game.state.PlayerInventory;
 import com.controllerface.bvge.geometry.MeshRegistry;
 import com.controllerface.bvge.geometry.ModelRegistry;
