+++ src/main/java/com/controllerface/bvge/game/world/WorldLoader.java
@@ -1,11 +1,10 @@
-package com.controllerface.bvge.ecs.systems.sectors;
+package com.controllerface.bvge.game.world;
 
 import com.controllerface.bvge.cl.GPGPU;
 import com.controllerface.bvge.ecs.ECS;
 import com.controllerface.bvge.ecs.systems.GameSystem;
+import com.controllerface.bvge.game.world.sectors.Sector;
 import com.controllerface.bvge.editor.Editor;
-import com.controllerface.bvge.game.world.EarthLikeWorld;
-import com.controllerface.bvge.game.world.WorldType;
 import com.controllerface.bvge.gl.renderers.UniformGridRenderer;
 import com.controllerface.bvge.physics.PhysicsEntityBatch;
 import com.controllerface.bvge.physics.UniformGrid;
