+++ src/main/java/com/controllerface/bvge/game/world/sectors/Sector.java
@@ -1,0 +1,3 @@
+package com.controllerface.bvge.game.world.sectors;
+
+public record Sector(int x, int y) { }
