+++ src/main/java/com/controllerface/bvge/game/world/sectors/SectorContainer.java
@@ -1,4 +1,4 @@
-package com.controllerface.bvge.ecs.systems.sectors;
+package com.controllerface.bvge.game.world.sectors;
 
 public interface SectorContainer
 {
