+++ src/main/java/com/controllerface/bvge/physics/PhysicsEntityBatch.java
@@ -1,6 +1,5 @@
 package com.controllerface.bvge.physics;
 
-import com.controllerface.bvge.ecs.systems.sectors.Sector;
 import com.controllerface.bvge.geometry.UnloadedEntity;
 import com.controllerface.bvge.substances.Solid;
